# IKM-WH/backend-node.js/swagger.yaml
openapi: 3.0.3
info:
  title: Festival Management System API
  description: |
    Полнофункциональный API для системы управления музыкальными фестивалями.
    
    ## Возможности
    - Управление заявками на участие в фестивале
    - Администрирование групп, мероприятий и фестивалей
    - Управление мерчендайзом и коллекционными карточками
    - Система карточек участников с загрузкой изображений
    - Управление афишами и постерами
    - Контент-менеджмент для сайта
    - Статистика и аналитика
    
    ## Аутентификация
    API использует JWT токены для аутентификации администраторов.
    Публичные эндпоинты доступны без аутентификации.
    
  version: 2.0.0
  contact:
    name: Festival Management System
    email: support@festival-system.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3001/api
    description: Development server
  - url: https://api.festival-system.com/api
    description: Production server
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT токен для аутентификации администраторов
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: string
          description: Описание ошибки
        details:
          type: object
          description: Дополнительные детали ошибки
...
paths:
...
  /uploads/{filename}:
    get:
      tags: [Utility]
      summary: Доступ к загруженным файлам
      description: Получение загруженных файлов (изображения, документы)
      parameters:
        - name: filename
          in: path
          required: true
          description: Имя файла
          schema:
            type: string
      responses:
        200:
          description: Файл найден
          content:
            image/*:
              schema:
                type: string
                format: binary
            application/octet-stream:
              schema:
                type: string
                format: binary
        404:
          description: Файл не найден
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

# Security definitions
security:
  - BearerAuth: []

# Examples and additional documentation
...